requirements.txt
django_admin_bootstrapped